Hackathone/backend/requirements.txt
fastapi
uvicorn
python-dotenv
cohere
qdrant-client
openai
pydantic
mangum
agents